+++ src/Products/PloneMeeting/browser/views_unrestricted.py
@@ -1,7 +1,10 @@
+# -*- coding: utf-8 -*-
+
 from AccessControl import Unauthorized
 from plone import api
 from plone.memoize.view import memoize
 from Products.Five import BrowserView
+from Products.PloneMeeting.browser.itemchangeorder import _is_integer
 from Products.PloneMeeting.utils import notifyModifiedAndReindex
 from zope.component import getMultiAdapter
 from zope.i18n import translate
@@ -32,8 +35,7 @@ class ItemUnrestrictedMethodsView(BrowserView):
 
 class MeetingUnrestrictedMethodsView(BrowserView):
 
-    @memoize
-    def findFirstItemNumber(self):
+    def findFirstItemNumber(self, get_items_additional_catalog_query={}):
         """
           Return the base number to take into account while computing an item number.
           This is used when given p_meeting firstItemNumber is -1, we need to look in previous
@@ -69,9 +71,13 @@ class MeetingUnrestrictedMethodsView(BrowserView):
             # and we continue to the previous meeting
             # divide lastItem itemNumber by 100 so we are sure to ignore subnumbers
             # 308 will become 3 or 1400 will become 14
-            items = meeting.get_items(the_objects=False, ordered=True, unrestricted=True)
-            item = items and items[-1]._unrestrictedGetObject()
-            lastItemNumber = item and item.getItemNumber() / 100 or 0
+            items = meeting.get_items(
+                the_objects=True,
+                ordered=True,
+                unrestricted=True,
+                additional_catalog_query=get_items_additional_catalog_query)
+            # compute number of items ignoring items with a subnumber
+            lastItemNumber = len([item for item in items if _is_integer(item.getItemNumber())])
             numberOfItemsBefore += lastItemNumber
             if not meeting.first_item_number == -1:
                 previousFirstItemNumber = meeting.first_item_number